Moved addBlock methods as well
