+++ src/main/java/control/SaveLoadController.java
@@ -2,7 +2,10 @@ package control;
 
 import java.io.File;
 
+import data.CameraShot;
+import data.CameraTimeline;
 import data.ScriptingProject;
+import gui.centerarea.CameraShotBlock;
 import gui.root.RootCenterArea;
 import gui.root.RootPane;
 import javafx.scene.control.Alert;
@@ -14,13 +17,11 @@ import lombok.extern.log4j.Log4j2;
 @Log4j2
 public class SaveLoadController {
     
-    private RootPane rootPane;
     
     private ControllerManager controllerManager;
     
     public SaveLoadController(ControllerManager manager) {
-        this.controllerManager = controllerManager;
-        this.rootPane = controllerManager.getRootPane();
+        this.controllerManager = manager;
     }
     
     /**
@@ -33,7 +34,7 @@ public class SaveLoadController {
         fileChooser.setTitle("Save");
         ExtensionFilter extFilter = new ExtensionFilter("txt files", "*.txt");
         fileChooser.getExtensionFilters().add(extFilter);
-        File file = fileChooser.showSaveDialog(rootPane.getPrimaryStage());
+        File file = fileChooser.showSaveDialog(controllerManager.getRootPane().getPrimaryStage());
         if (file != null) {
             controllerManager.getScriptingProject().write(file);
         }
@@ -49,7 +50,7 @@ public class SaveLoadController {
         fileChooser.setTitle("Load");
         ExtensionFilter extFilter = new ExtensionFilter("txt files", "*.txt");
         fileChooser.getExtensionFilters().add(extFilter);
-        File file = fileChooser.showOpenDialog(rootPane.getPrimaryStage());
+        File file = fileChooser.showOpenDialog(controllerManager.getRootPane().getPrimaryStage());
         if (file != null) {
             ScriptingProject temp  = ScriptingProject.read(file);
             if (temp == null) {
@@ -59,11 +60,40 @@ public class SaveLoadController {
                 alert.showAndWait();
             } else {
                 controllerManager.setScriptingProject(temp);
-                rootPane.reInitRootCenterArea(new RootCenterArea(rootPane, controllerManager.getScriptingProject().getCameraTimelines().size(), false));
-                controllerManager.getTimelineControl().addLoadedBlocks(controllerManager.getScriptingProject());
+                controllerManager.getRootPane().reInitRootCenterArea(new RootCenterArea(controllerManager.getRootPane(), controllerManager.getScriptingProject().getCameraTimelines().size(), false));
+                addLoadedBlocks(controllerManager.getScriptingProject());
                 controllerManager.getTimelineControl().setNumTimelines(controllerManager.getScriptingProject().getCameraTimelines().size());
             }
         }
+    }
+    
+    /**
+     * Add the blocks that were loaded from file to the UI.
+     * @param project the project that was loaded
+     */
+    public void addLoadedBlocks(ScriptingProject project) {
+        log.info("Adding loaded blocks");
+        for (int i = 0; i < project.getCameraTimelines().size();i++) {
+            CameraTimeline timeline = project.getCameraTimelines().get(i);
+            int amountShots = timeline.getShots().size();
+            for (int j = 0; j < amountShots;j++) {
+                CameraShot shot = timeline.getShots().get(j);
+                addCameraShotForLoad(i, shot);
+            }
+        }
+    }
+    
+    /**
+     * Add a camera shot that is loaded from file.
+     * @param cameraIndex the index of the camera timeline to use
+     * @param shot the shot to add
+     */
+    private void addCameraShotForLoad(int cameraIndex, CameraShot shot) {
+        CameraShotBlock shotBlock = new CameraShotBlock(shot, cameraIndex, 
+                                                       controllerManager.getRootPane().getRootCenterArea(),
+                                                       controllerManager.getTimelineControl()::shotChangedHandler);
+        controllerManager.setActiveShotBlock(shotBlock);
+        controllerManager.getTimelineControl().getCameraShotBlocks().add(shotBlock);
     }
 
 }
